refactor: load Opal-specific NativeAttachment via runtime compatibility
